backend/requirements.txt
# 백엔드 필수 패키지 목록 (가상환경에서 설치)
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
flasgger==0.9.7.1
Pillow==10.3.0
swagger==2.0
